Fixes #2256. Renames LinuxTime into PosixTime implementing new Time interface
