+++ Svc/CMakeLists.txt
@@ -40,7 +40,6 @@ add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/PolyDb/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/PrmDb/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/RateGroupDriver/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/StaticMemory/")
-add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Time/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/TlmChan/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/TlmPacketizer/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/SystemResources/")
@@ -53,6 +52,6 @@ if (FPRIME_ENABLE_TEXT_LOGGERS)
 endif()
 
 if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux" OR ${CMAKE_SYSTEM_NAME} STREQUAL "Darwin")
-	add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LinuxTime/")
+	add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/PosixTime/")
 	add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LinuxTimer/")
 endif()
