--- Svc/LinuxTime/docs/sdd.md
@@ -1,35 +1,0 @@
-\page SvcLinuxTimeComponent Svc::LinuxTime Component
-# Svc::LinuxTime Component
-
-## 1. Introduction
-
-The `Svc::LinuxTime` is a component that provides system time on a Linux system.
-
-## 2. Requirements
-
-TBD
-
-## 3. Design
-
-### 3.1 Context
-
-#### 3.1.1 Component Diagram
-
-TBD
-
-## 4. Dictionaries
-
-Not applicable
-
-## 5. Module Checklists
-
-## 6. Unit Testing
-
-## 7. Change Log
-
-Date | Description
----- | -----------
-4/20/2017 | Initial Version
-
-
-
