+++ Svc/PosixTime/PosixTime.fpp
@@ -1,6 +1,6 @@
 module Svc {
   @ A component for getting time
-  passive component LinuxTime {
+  passive component PosixTime {
     include "../Interfaces/Time.fppi"
   }
 }
