--- Svc/Time/CMakeLists.txt
@@ -1,13 +1,0 @@
-####
-# F prime CMakeLists.txt:
-#
-# SOURCE_FILES: combined list of source and autocoding files
-# MOD_DEPS: (optional) module dependencies
-#
-# Note: using PROJECT_NAME as EXECUTABLE_NAME
-####
-set(SOURCE_FILES
-  "${CMAKE_CURRENT_LIST_DIR}/Time.fpp"
-)
-
-register_fprime_module()
