--- Svc/Time/docs/.gitignore
@@ -1,1 +1,0 @@
-*.html
